[DEV] Organised package structure.
