+++ Tests/SavingDataTests.java
@@ -1,8 +1,6 @@
-import data.Storage;
-import data.Token;
+import data.security.Token;
 import org.junit.jupiter.api.Test;
-import models.Account;
-import data.Database;
+import data.database.Database;
 import models.User;
 
 import java.io.IOException;
