+++ src/main/java/controllers/CodeViewControllers/CodeViewController.java
@@ -15,7 +15,7 @@ import javafx.scene.input.KeyEvent;
 import javafx.scene.layout.*;
 import javafx.stage.FileChooser;
 import javafx.scene.input.KeyCode;
-import data.Database;
+import data.database.Database;
 import models.Account;
 import views.*;
 import views.interfaces.Reversible;
