+++ src/main/java/controllers/SignInController.java
@@ -4,9 +4,9 @@ import commands.Command;
 import commands.SwitchToDarkMode;
 import commands.SwitchToHighContrastMode;
 import commands.managers.ThemeSwitcher;
-import data.Database;
+import data.database.Database;
 import data.Storage;
-import data.Token;
+import data.security.Token;
 import javafx.beans.binding.Bindings;
 import javafx.beans.property.DoubleProperty;
 import javafx.beans.property.SimpleDoubleProperty;
